requirements.txt
python-telegram-bot==20.7
pandas==2.0.3
numpy==1.24.3